+++ keel-engine-kqueue/src/macosMain/kotlin/io/github/fukusaka/keel/engine/kqueue/KqueueServer.kt
@@ -90,9 +90,7 @@ internal class KqueueServer(
         while (true) {
             val clientFd = accept(serverFd, null, null)
             if (clientFd >= 0) {
-                PosixSocketUtils.setNonBlocking(clientFd)
-                val remoteAddr = PosixSocketUtils.getRemoteAddress(clientFd)
-                val localAddr = PosixSocketUtils.getLocalAddress(clientFd)
+                val (remoteAddr, localAddr) = PosixSocketUtils.acceptClient(clientFd)
                 val (workerLoop, allocator) = workerGroup.next()
                 val transport = KqueueIoTransport(clientFd, workerLoop, allocator)
                 val channel = KqueuePipelinedChannel(